MovableObjects and LockedBoxes
Added the abstract MovableObject class for objects that can be picked up and moved. From this, added cheese for the mouse, a Key to be used to unlock a box. Secondly, added the abstract LockedBoxed class for both a locked box to be opened by the key and one to be opened by the book code. Implemented the key box but not the book one.